pyproject.toml
[project]
name = "graphite-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arize-phoenix>=11.28.0",
    "arize-phoenix-client>=1.15.3",
    "fastmcp>=2.10.6",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "googlesearch-python>=1.3.0",
    "grafi==0.0.25",
    "grafi-dev>=0.0.7",
    "jinja2>=3.1.6",
    "mypy>=1.17.0",
    "pandas>=2.3.1",
    "psycopg2>=2.9.10",
    "py-solc-x>=2.0.4",
    "pycountry>=24.6.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "ruff>=0.12.5",
    "sqlalchemy>=2.0.42",
    "web3>=7.12.1",
]
